@@ -1,9 +1,10 @@
 import './App.css';
+import Homepage from "./Homepage";
 
 function App() {
   return (
     <div className="App">
-      <h1>Henry Dogs</h1>
+      <Homepage />
     </div>
   );
 }
